Move to static GTFS files
An issue arose where Heroku was unable to process the ZIP file
containing GTFS data within the timeout limit of the application server.
In reality, the update process added in this PR should write to a data
store (RDBMS or cache service), but I did not want to introduce another
component to the stack. The total app weigns in around 200MB, which is
less than the limit Heroku imposes.