Add tire tracks to show bearing, highlight projected stop
This PR adds a set of tire tracks behind the bus marker that is rotated to
show which way, at a glance, the bus is heading. Required adding a
modified plugin to support rotating the shadow of a marker independent
of the marker itself.

Also highlights the projected position in the stop sequence based on the
current time when viewing the trip details, as well as showing when the
trip update was last posted.