Cache realtime data locally
Fixes #5

Prior to this PR, every client that connected to the application would
result in a hit to the API every 10 seconds, multiplied out by however
many users are on the site. This change introduces a local file cache
that will return that cached version for every request within the same
ten second window. In effect, we should never be hitting the API more
than six times a minute, no matter the load on the site.