Add agency information, add RTA bus icon
Nashville is a bit unique in that it has more than one "agency" for its
transit system, the MTA (WeGo) and the RTA (maybe also WeGo?). Either
way, this PR shows the relevant agency responsible for a particular
route/bus and shows a different icon if it's RTA.

Also refactors the route listing to be part of the initial page load
rather than retrieved over AJAX. This greatly simplifies the JavaScript
and allows for certain things to skip the network request to get
additional information.