New about page and base map layer
Switched out the base map layer for one from CartoDB that is a bit
lighter and clashes less with the route and bus overlays. Also added an
"About" page to describe the project and what technology is used in it.
Fixes a bug where the Google Analytics profile ID was incorrect.