Add a location button to UI
This PR adds a location button that will appear on devices where it is
possible to leverage GPS or network data to provide a latitude and
longitude of the user if granted. It prevents a user from outside of the
Middle Tennessee area (governed by the map's maxBounds) to use this
feature, as it would strand their view at the edge of the map.

Also adds a requirement that a production deploy must use SSL, as it is
a requirement for the above.